adjusted UIManager classes & initial implementation of main screen
I realised it's probably a better idea to have an intermediary class (like a controller?) to contain logic for button presses. This way, for example, instead of screens being responsible for navigating to other screens, we can do:

button is pressed -> function called on controller -> controller uses UI manager to navigate to different screen

on top of this, we can use it to further separate logic from UI classes.